[172] Test, Query and Serivce update
